--- azure-communication-ui/azure-communication-ui-demo-app/src/main/res/drawable/memory_indicator.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-   Copyright (c) Microsoft Corporation. All rights reserved.
-   Licensed under the MIT License.
-  -->
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="oval">
-    <corners android:radius="25dip"/>
-    <size
-        android:width="50dp"
-        android:height="50dp" />
-    <stroke android:color="@color/colorPrimary" android:width="1dp"/>
-    <solid android:color="@android:color/white"/>
-</shape>
