+++ azure-communication-ui/azure-communication-ui-demo-app/src/main/res/layout/memory_view.xml
@@ -6,10 +6,14 @@
 <TextView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:background="@drawable/memory_indicator"
     android:gravity="center"
     android:padding="10dp"
-    android:text="60"
-    android:textColor="@android:color/black"
+    android:clipChildren="false"
+    android:clipToPadding="false"
+    android:shadowDx="10"
+    android:shadowDy="10"
+    android:shadowRadius="20"
+    android:shadowColor="@color/colorPrimary"
+    android:textColor="@color/colorPrimary"
     android:textSize="18dp"
-    android:visibility="visible" />
+    android:visibility="gone" />
